v0.2

+++ Customer.java
@@ -1,8 +1,10 @@
 package com.company;
 
+
 public class Customer {
+
 	//datafield tiplerini degistirebilirsiniz
-	int arrivalTime; //musteri gelis zamani-
+	double arrivalTime; //musteri gelis zamani-
 	//bekleme zamanini hesaplamada kullanabilirsiniz
-	int removalTime;
+	double removalTime;
 }
